// frontend/src/app/models/registro.table.model.ts
export interface RegistroTable {
    orderkey:        string;
    name:            string;
    email:           string;
    shippingCity:    string;
    shippingAdress1: string;
    shippingAdress2: string;
    descStatus:      string;
    cargaDt:         Date;
    scheduledDt:     Date;
    comment:         string;
    courier:         string;
    vendedor:        string;
}